Configuration files and wsgi files for Docker Passenger
This is a major commit of the Dockerfile used to build the Docker image
used in production. The Dockerrun file is used by Elastic Beanstalk
to pull down the image and dockercfg files from an S3 bucket. The
passenger_wsgi.py file is used by the Passenger webserver to proxy Nginx
in production. It is the WSGI source for the Passenger webserver.